move RegionalPartnerPlaybook component out of 'entries' directory and into generic ui component directory
